+++ Cargo.toml
@@ -21,7 +21,7 @@ async-trait = "0.1.53"
 log = "0.4.17"
 env_logger = "0.9.0"
 hex = "0.4.3"
-tokio = { version = "1.19.2", features = ["macros", "rt-multi-thread"] }
+tokio = { version = "1.19.2", features = ["macros", "rt-multi-thread", "fs"] }
 codec = { version = "3.0.0", package = "parity-scale-codec" }
 clap = { version = "3.2.22", features = ["derive"] }
 toml = "0.5.9"
